Refactor olm pkg list functions
